Integrate authentication features: add LoginForm and RegisterForm components, implement AuthContext for user management, and refactor ProfileScreen to utilize new authentication flow. Update axios instance for API requests and adjust placeService imports.
